# api/openapi.yaml
openapi: 3.1.0
info:
  title: INFRA-AI investor API
  version: 0.1.0
  summary: Explainable infrastructure scores for the cities and areas of India.
  description: |
    Read-only JSON API behind the "where should I invest?" flow. A score is a config-weighted sum of observed
    infrastructure facts with stored drivers (ADR-0003). *Scores describe existing infrastructure and past
    growth. They are not forecasts, price predictions or financial advice.*

    **Data.** Loaded once at start-up from the fixtures the pipeline exports to `web/fixtures/invest/`
    (design spec section 6); the Go store holds no scoring logic. Every example in this file comes from the
    synthetic test world in `api/testdata/invest/` (each record has `source: synthetic-test`), so the numbers
    are invented and must never be shown as real.

    **Errors.** Always `{"error": {"code", "message"}}`. The message is safe to show and never carries
    internal detail. Codes: `bad_request` 400, `not_found` 404, `rate_limited` 429 (with `Retry-After`),
    `timeout` 503 (the request outlived its 5 s budget) and `internal` 500.

    **Validation.** State codes match `^[A-Z]{2}$` and city ids `^[a-z0-9-]{2,64}$`; both are looked up in
    memory and never used to build a path. An unknown `preset`, an unknown enum value or an out-of-range
    `limit` is a 400 - a limit is never clamped silently.

    **Empty and unknown.** A list is never `null`: an empty list is `[]`. A `null` inside a record means "not
...

...
servers:
...
paths:
...
  /v1/cities:
    get:
      operationId: searchCities
      tags: [cities]
      summary: Find a city by name or alias
      description: |
        Case-insensitive match on city names and aliases (an alias is another big place inside the same urban
        centre, so "Thane" finds Mumbai); names that start with the text come before names that merely contain
        it. `matched` is the name or alias that matched. No match is a 200 with an empty list.
      parameters:
        - name: q
          in: query
          required: true
          description: Search text of 2-64 characters, in any script. Too short or too long is a 400. Never echoed back unescaped.
          schema: {type: string, minLength: 2, maxLength: 64}
          example: thane
        - name: limit
          in: query
          description: Maximum number of matches in the response. Outside 1-20 is a 400.
          schema: {type: integer, minimum: 1, maximum: 20, default: 8}
      responses:
        "200":
          description: Matching cities, best match first.
          headers:
            Cache-Control: {$ref: "#/components/headers/CacheControl"}
          content:
            application/json:
              schema: {$ref: "#/components/schemas/SearchResponse"}
              examples:
                alias:
                  summary: q=thane finds Mumbai through its alias
                  value:
                    cities:
                      - {id: mumbai, name: Mumbai, state: MH, tier: metro, population: 12600000, matched: Thane}
                nothing:
                  summary: q=zzz
                  value:
                    cities: []
        "400": {$ref: "#/components/responses/BadRequest"}
        "429": {$ref: "#/components/responses/RateLimited"}
        "500": {$ref: "#/components/responses/Internal"}
        "503": {$ref: "#/components/responses/Timeout"}
...
  /v1/cities/{id}/compare:
    get:
      operationId: compareCity
      tags: [cities]
      summary: The base city against other cities, with the reasons for each difference
      description: |
        `others` never contains the base city and is ordered by score, best first, so better and worse cities
        both appear. `delta` is the other city's score minus the base city's, as served. `better` and `worse`
        hold up to two sub-score differences each (largest first, ties by factor id): where the other city is
        stronger and where it is weaker - the reasons behind `delta`. `base_rank` is the base city's rank in
        the compared set including itself and `total` the number of other cities in scope, before `limit`.
        The default scope is `metros` when the base city is a metro and `peers` otherwise; a city with no
        peers gets `others: []` and `total: 0`.
      parameters:
        - $ref: "#/components/parameters/CityId"
        - $ref: "#/components/parameters/Preset"
        - name: scope
          in: query
          description: |
            Cities to compare against: `metros` the other metros, `peers` cities of the same tier across India,
            `state` the other cities of the base city's state, `india` every other city.
          schema: {type: string, enum: [metros, peers, state, india]}
        - name: limit
          in: query
          description: Maximum number of other cities in the response. Outside 1-10 is a 400.
          schema: {type: integer, minimum: 1, maximum: 10, default: 5}
      responses:
        "200":
          description: The comparison.
          headers:
            Cache-Control: {$ref: "#/components/headers/CacheControl"}
          content:
            application/json:
              schema: {$ref: "#/components/schemas/CompareResponse"}
              examples:
                pune_commuter_metros:
                  summary: Pune, preset=commuter, default scope (metros)
                  value:
                    base:
                      rank: 3
                      id: pune
                      name: Pune
                      state: MH
                      tier: metro
                      population: 6100000
                      score: 63.5
                      access: 54.7
                      momentum: 90.0
                      coverage: 1.0
                      confidence: 0.8
                      drivers:
                        - {factor: built_up_growth, points: 22.5, value: 28.0, unit: pp, share: null, band_km: null}
                        - {factor: rail_access, points: 15.0, value: 3.6, unit: km, share: 0.0, band_km: 2}
                        - {factor: nh_access, points: 7.5, value: 4.3, unit: km, share: 1.0, band_km: 10}
                      gaps: []
                      best_area: {id: 876088500ffffff, name: Hinjewadi, score: 70.0}
                      source: synthetic-test
                      source_ref: city:pune
                      fetched_at: '2026-09-29'
                      license: synthetic-test
                    preset: commuter
                    scope: metros
                    total: 3
                    base_rank: 3
                    others:
                      - rank: 1
                        id: delhi
                        name: Delhi
                        state: DL
                        tier: metro
                        score: 73.8
                        delta: 10.3
                        better:
                          - {factor: metro_access, delta: 50.0, base: 40.0, other: 90.0}
                          - {factor: rail_access, delta: 20.0, base: 60.0, other: 80.0}
                        worse:
                          - {factor: built_up_growth, delta: -45.0, base: 90.0, other: 45.0}
                          - {factor: nh_access, delta: -5.0, base: 75.0, other: 70.0}
                      - rank: 2
                        id: mumbai
                        name: Mumbai
                        state: MH
                        tier: metro
                        score: 69.5
                        delta: 6.0
                        better:
                          - {factor: metro_access, delta: 40.0, base: 40.0, other: 80.0}
                          - {factor: rail_access, delta: 30.0, base: 60.0, other: 90.0}
                        worse:
                          - {factor: built_up_growth, delta: -50.0, base: 90.0, other: 40.0}
                          - {factor: nh_access, delta: -15.0, base: 75.0, other: 60.0}
                      - rank: 4
                        id: bengaluru
                        name: Bengaluru
                        state: KA
                        tier: metro
                        score: 61.8
                        delta: -1.7
                        better:
                          - {factor: metro_access, delta: 20.0, base: 40.0, other: 60.0}
                        worse:
                          - {factor: rail_access, delta: -15.0, base: 60.0, other: 45.0}
                          - {factor: built_up_growth, delta: -10.0, base: 90.0, other: 80.0}
                nashik_alone:
                  summary: Nashik is the only large city, so its default scope (peers) is empty
                  value:
                    base:
                      rank: 1
                      id: nashik
                      name: Nashik
                      state: MH
                      tier: large
                      population: 1900000
                      score: 55.0
                      access: 48.6
                      momentum: 70.0
                      coverage: 1.0
                      confidence: 0.8
                      drivers:
                        - {factor: nh_access, points: 21.3, value: 2.8, unit: km, share: 1.0, band_km: 10}
                        - {factor: built_up_growth, points: 21.0, value: 15.0, unit: pp, share: null, band_km: null}
                      gaps:
                        - {factor: metro_access, subscore: 0.0, value: 167.9, unit: km, share: 0.0, band_km: 2}
                      best_area: {id: 876099a04ffffff, name: Gangapur, score: 55.0}
                      source: synthetic-test
                      source_ref: city:nashik
                      fetched_at: '2026-09-29'
                      license: synthetic-test
                    preset: balanced
                    scope: peers
                    total: 0
                    base_rank: 1
                    others: []
        "400": {$ref: "#/components/responses/BadRequest"}
        "404": {$ref: "#/components/responses/NotFound"}
        "429": {$ref: "#/components/responses/RateLimited"}
        "500": {$ref: "#/components/responses/Internal"}
        "503": {$ref: "#/components/responses/Timeout"}
components:
  parameters:
...
    CityId:
      name: id
      in: path
      required: true
      description: City id, a slug. Malformed is a 400, unknown a 404.
      schema: {type: string, pattern: "^[a-z0-9-]{2,64}$"}
      example: pune
    Preset:
      name: preset
      in: query
      description: |
        Scoring preset; one of `presets[].id` in `/v1/meta`. Omitted means the preset marked `default`
        (`balanced`). Anything else is a 400.
      schema:
        $ref: "#/components/schemas/PresetId"
        default: balanced
  headers:
    CacheControl:
      description: Data routes may be cached for five minutes.
      schema: {type: string, const: "public, max-age=300"}
    RetryAfter:
      description: Seconds until the rate-limit window resets.
      schema: {type: integer, minimum: 0}
  responses:
    BadRequest:
      description: A path value, query value or limit is malformed or out of range.
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Error"}
          examples:
            limit:
              value: {error: {code: bad_request, message: "limit must be an integer from 1 to 20"}}
    NotFound:
      description: The state code or city id is well formed but not in the data.
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Error"}
          examples:
            city:
              value: {error: {code: not_found, message: "city not found"}}
    RateLimited:
      description: This client exceeded the per-minute request limit.
      headers:
        Retry-After: {$ref: "#/components/headers/RetryAfter"}
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Error"}
          examples:
            limited:
              value: {error: {code: rate_limited, message: "too many requests; retry later"}}
    Internal:
      description: An unexpected failure. Nothing internal is disclosed.
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Error"}
          examples:
            failure:
              value: {error: {code: internal, message: "internal error"}}
    Timeout:
      description: The request took longer than its time budget (5 s) and was given up. Retrying is safe, because every route only reads.
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Error"}
          examples:
            slow:
              value: {error: {code: timeout, message: "the request took too long"}}
  schemas:
    Error:
      type: object
      required: [error]
      additionalProperties: false
      properties:
        error:
          type: object
          required: [code, message]
          additionalProperties: false
          properties:
            code: {type: string, enum: [bad_request, not_found, rate_limited, timeout, internal]}
            message: {type: string, description: Human-readable and safe to show.}
...
    PresetId:
      type: string
      enum: [balanced, commuter, highway, growth]
    Tier:
      type: string
      enum: [metro, large, mid]
...
    CityBase:
      type: object
      description: What a city looks like for one preset - the ranked card without its rank.
      required: [id, name, state, tier, population, score, access, momentum, coverage, confidence, drivers, gaps, best_area, source, source_ref, fetched_at, license]
      properties:
        id: {type: string, pattern: "^[a-z0-9-]{2,64}$"}
        name: {type: string}
        state: {type: string, pattern: "^[A-Z]{2}$", description: ISO state code; shown next to the name because names repeat across states.}
        tier: {$ref: "#/components/schemas/Tier"}
        population: {type: integer, description: People in the urban centre inside the state (WorldPop 2020).}
        score: {type: number, minimum: 0, maximum: 100, description: Population-weighted mean of the city's area scores.}
        access: {type: [number, "null"], minimum: 0, maximum: 100, description: The same mean over the access factors only.}
        momentum: {type: [number, "null"], minimum: 0, maximum: 100, description: The built-up growth sub-score.}
        coverage: {type: number, minimum: 0, maximum: 1, description: Share of the preset's weight backed by observed data.}
        confidence: {type: number, minimum: 0, maximum: 1, description: 0.8 x coverage; 0.8 reflects community and modelled sources.}
        drivers: {type: array, maxItems: 3, items: {$ref: "#/components/schemas/Driver"}}
        gaps: {type: array, maxItems: 2, items: {$ref: "#/components/schemas/Gap"}}
        best_area: {$ref: "#/components/schemas/BestArea"}
        source: {type: string}
        source_ref: {type: string}
        fetched_at: {type: string, format: date}
        license: {type: string}
    CityCard:
      description: A city in a ranked list.
      allOf:
        - $ref: "#/components/schemas/CityBase"
        - type: object
          required: [rank]
          properties:
            rank: {type: integer, minimum: 1, description: "Position in the list it appears in (in a comparison, in the compared set including the base city)."}
      unevaluatedProperties: false
...
    SearchHit:
      type: object
      required: [id, name, state, tier, population, matched]
      additionalProperties: false
      properties:
        id: {type: string}
        name: {type: string}
        state: {type: string, pattern: "^[A-Z]{2}$"}
        tier: {$ref: "#/components/schemas/Tier"}
        population: {type: integer}
        matched: {type: string, description: "The name or alias that matched, as stored."}
    SearchResponse:
      type: object
      required: [cities]
      additionalProperties: false
      properties:
        cities: {type: array, items: {$ref: "#/components/schemas/SearchHit"}}

    # ---- areas ----------------------------------------------------------------------------------------
...
    CompareRow:
      type: object
      required: [rank, id, name, state, tier, score, delta, better, worse]
      additionalProperties: false
      properties:
        rank: {type: integer, minimum: 1, description: Position in the compared set including the base city.}
        id: {type: string}
        name: {type: string}
        state: {type: string, pattern: "^[A-Z]{2}$"}
        tier: {$ref: "#/components/schemas/Tier"}
        score: {type: number, minimum: 0, maximum: 100}
        delta: {type: number, description: "This city's score minus the base city's, one decimal."}
        better: {type: array, maxItems: 2, items: {$ref: "#/components/schemas/FactorDelta"}, description: Where this city is stronger (delta above 0).}
        worse: {type: array, maxItems: 2, items: {$ref: "#/components/schemas/FactorDelta"}, description: Where this city is weaker (delta below 0).}
    CompareResponse:
      type: object
      required: [base, preset, scope, total, base_rank, others]
      additionalProperties: false
      properties:
        base: {$ref: "#/components/schemas/CityCard"}
        preset: {$ref: "#/components/schemas/PresetId"}
        scope: {type: string, enum: [metros, peers, state, india]}
        total: {type: integer, minimum: 0, description: "Other cities in scope, before `limit`."}
        base_rank: {type: integer, minimum: 1, description: The base city's rank in the compared set including itself (also `base.rank`).}
        others: {type: array, items: {$ref: "#/components/schemas/CompareRow"}}